fix to hide unneccesary stuff when payson is alone
This it not a optimal fix, but it works. I would just like to point out why it isn't working as default. 

Maybe you know if there's some JS event called when the checkout form is fully initiated. Then you should hook up into that event instead of this ugly setInterval solution which is horrible.