Update nonces in case the customer has been signed in
Since the nonce is associated with the user ID, when a guest user (ID = 0) is signed in during the checkout, the nonce will become invalid, resulting in all AJAX requests failing validation. For this reason, we need to update the nonce, and since signing in is handled by Woo when `process_payment` is triggered, we send the nonce to the client in the same function.